I turn all doubles to floats

+++ lib/pid/include/PID.h
@@ -63,8 +63,8 @@
 class PID {
  public:
   typedef struct {
-    double P, I, D, bias = 0.0;
-    double deadZoneError = 0.0;
+    float P, I, D, bias = 0.0;
+    float deadZoneError = 0.0;
   } t_pidConfig;
 
   // add other actuators here
@@ -73,8 +73,8 @@ class PID {
   // configurable parameters for autotuning process
   typedef struct {
     int nLookBack;
-    std::chrono::duration<double> sampleTime;
-    double outputStart, oStep, noiseBand, setpoint;
+    std::chrono::duration<float> sampleTime;
+    float outputStart, oStep, noiseBand, setpoint;
   } t_AutoTuneConfig;
 
   /**
@@ -88,7 +88,7 @@ class PID {
    * @param tauD - Tuning parameter
    * @param interval PID calculation performed every interval seconds.
    */
-  PID(double Kc, double tauI, double tauD, std::chrono::duration<double> interval);
+  PID(float Kc, float tauI, float tauD, std::chrono::duration<float> interval);
 
   /**
    * Scale from inputs to 0-100%.
@@ -96,7 +96,7 @@ class PID {
    * @param InMin The real world value corresponding to 0%.
    * @param InMax The real world value corresponding to 100%.
    */
-  void setInputLimits(double inMin, double inMax);
+  void setInputLimits(float inMin, float inMax);
 
   /**
    * Scale from outputs to 0-100%.
@@ -104,7 +104,7 @@ class PID {
    * @param outMin The real world value corresponding to 0%.
    * @param outMax The real world value corresponding to 100%.
    */
-  void setOutputLimits(double outMin, double outMax);
+  void setOutputLimits(float outMin, float outMax);
 
   /**
    * Calculate PID constants.
@@ -115,7 +115,7 @@ class PID {
    * @param tauI - Tuning parameter
    * @param tauD - Tuning parameter
    */
-  void setTunings(double Kc, double tauI, double tauD);
+  void setTunings(float Kc, float tauI, float tauD);
 
   /**
    * Reinitializes controller internals. Automatically
@@ -136,40 +136,40 @@ class PID {
    *
    * @param interval PID calculation peformed every interval seconds.
    */
-  void setInterval(std::chrono::duration<double> interval);
+  void setInterval(std::chrono::duration<float> interval);
 
   /**
    * Set the set point.
    *
    * @param sp The set point as a real world value.
    */
-  void setSetPoint(double sp);
+  void setSetPoint(float sp);
 
   /**
    * Set the process value.
    *
    * @param pv The process value as a real world value.
    */
-  void setProcessValue(double pv);
+  void setProcessValue(float pv);
 
   /**
    * Set the bias.
    *
    * @param bias The bias for the controller output.
    */
-  void setBias(double bias);
+  void setBias(float bias);
 
   /**
    * Set the dead zone error to allow error to round down if within +/- this value
    * @param error Round error down to 0.0 if error is within +/- this value
    */
-  void setDeadZoneError(double error);
+  void setDeadZoneError(float error);
 
   /**
    * Set real output value
    * @param realOutput real life output
    */
-  void setRealOutput(double realOutput);
+  void setRealOutput(float realOutput);
 
   /**
    * Setup needed before autotuning
@@ -177,7 +177,7 @@ class PID {
    * @param inputPointer pointer to calculated input value, usually done in background on main
    * @param actuatorType type of actuator needed to cast generic void pointer
    */
-  void setupAutoTune(void *outputPointer, double *inputPointer, int actuatorType);
+  void setupAutoTune(void *outputPointer, float *inputPointer, int actuatorType);
 
   /**
    * Run the autotuning algorithm and set autoTune class member variables
@@ -192,7 +192,7 @@ class PID {
    * Set the output from within the PID class, needed for autotuning
    * @param output output as a double, casting done in function based on actuator type
    */
-  void setOutput(double output);
+  void setOutput(float output);
 
   /**
    * Set the tuning values to the calculated autotuning parameters
@@ -204,39 +204,39 @@ class PID {
    *
    * @return The controller output as a double between outMin and outMax.
    */
-  double compute(void);
+  float compute(void);
 
   // Getters.
-  double getInMin();
-  double getInMax();
-  double getOutMin();
-  double getOutMax();
-  std::chrono::duration<double> getInterval();
-  double getPParam();
-  double getIParam();
-  double getDParam();
-  double getATunePParam();
-  double getATuneIParam();
-  double getATuneDParam();
-  double getSetPoint();
+  float getInMin();
+  float getInMax();
+  float getOutMin();
+  float getOutMax();
+  std::chrono::duration<float> getInterval();
+  float getPParam();
+  float getIParam();
+  float getDParam();
+  float getATunePParam();
+  float getATuneIParam();
+  float getATuneDParam();
+  float getSetPoint();
 
  private:
   bool usingFeedForward;
   bool inAuto;
 
   // Actual tuning parameters used in PID calculation.
-  double Kc_;
-  double tauR_;
-  double tauD_;
+  float Kc_;
+  float tauR_;
+  float tauD_;
 
   // Tuning parameters calculated from autotuning function
-  double autoTuneKc_;
-  double autoTuneTauR_;
-  double autoTuneTauD_;
+  float autoTuneKc_;
+  float autoTuneTauR_;
+  float autoTuneTauD_;
 
   // Input/Output pointers for AutoTuning
   void *output_;
-  double *input_;
+  float *input_;
 
   // Store type of actuator PID is being used on
   int actuatorType_;
@@ -245,38 +245,38 @@ class PID {
   t_AutoTuneConfig AutoTuneConfig;
 
   // Raw tuning parameters.
-  double pParam_;
-  double iParam_;
-  double dParam_;
+  float pParam_;
+  float iParam_;
+  float dParam_;
 
   // The point we want to reach.
-  double setPoint_;
+  float setPoint_;
   // The thing we measure.
-  double processVariable_;
-  double prevProcessVariable_;
+  float processVariable_;
+  float prevProcessVariable_;
   // The output that affects the process variable.
-  double controllerOutput_;
-  double prevControllerOutput_;
+  float controllerOutput_;
+  float prevControllerOutput_;
 
   // We work in % for calculations so these will scale from
   // real world values to 0-100% and back again.
-  double inMin_;
-  double inMax_;
-  double inSpan_;
-  double outMin_;
-  double outMax_;
-  double outSpan_;
+  float inMin_;
+  float inMax_;
+  float inSpan_;
+  float outMin_;
+  float outMax_;
+  float outSpan_;
 
   // The accumulated error, i.e. integral.
-  double accError_;
+  float accError_;
   // The allowed error range for error to be rounded to 0.0
-  double deadZoneError_;
+  float deadZoneError_;
   // The controller output bias.
-  double bias_;
+  float bias_;
 
   // The interval between samples.
-  std::chrono::duration<double> tSample_;
+  std::chrono::duration<float> tSample_;
 
   // Controller output as a real world value.
-  volatile double realOutput_;
+  volatile float realOutput_;
 };
